Creates a beastier report

Creates a beastier report, to be used when reporting bugs. Uses message

Usage

beastier_report(
  beast2_folder = get_default_beast2_folder(),
  os = rappdirs::app_dir()$os
)

Arguments

beast2_folder

the folder where the BEAST2 is installed. Note that this is not the folder where the BEAST2 executable is installed: the BEAST2 executable is in a subfolder. Use get_default_beast2_folder to get the default BEAST2 folder. Use get_default_beast2_bin_path to get the full path to the default BEAST2 executable.

os

name of the operating system, must be unix (Linux, Mac) or win (Windows)

Value

No return value, the information will be shown using message

Author

Richèl J.C. Bilderbeek

Examples

check_empty_beaustier_folders()

beastier_report()
#> ***********
#> * beastier *
#> ***********
#> OS: unix
#> beast2_folder: ~/.local/share
#> beast2_path: /github/home/.local/share/beast/lib/launcher.jar
#> ****************
#> * Dependencies *
#> ****************
#> beautier version: 2.6.11
#> beastier version: 2.5
#> **********
#> * BEAST2 *
#> **********
#> Java version: 11.0.22+7-post-Ubuntu-0ubuntu222.04.1
#> Is BEAST2 installed: FALSE
#> ****************
#> * session info *
#> ****************
#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value
#>  version  R version 4.3.2 (2023-10-31)
#>  os       Ubuntu 22.04.3 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language en
#>  collate  C
#>  ctype    en_US.UTF-8
#>  tz       Etc/UTC
#>  date     2024-03-06
#>  pandoc   2.16.2 @ /usr/local/bin/ (via rmarkdown)
#> 
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version date (UTC) lib source
#>  beastier    * 2.5     2024-03-06 [1] local (/tmp/Rtmp22h0pn/beastier-source)
#>  beautier      2.6.11  2024-03-02 [1] https://ropensci.r-universe.dev (R 4.3.2)
#>  bslib         0.6.1   2023-11-28 [1] RSPM (R 4.3.0)
#>  cachem        1.0.8   2023-05-01 [1] RSPM (R 4.3.0)
#>  cli           3.6.2   2023-12-11 [1] RSPM (R 4.3.0)
#>  curl          5.2.0   2023-12-08 [1] RSPM (R 4.3.0)
#>  desc          1.4.3   2023-12-10 [1] RSPM (R 4.3.0)
#>  digest        0.6.34  2024-01-11 [1] RSPM (R 4.3.0)
#>  downlit       0.4.3   2023-06-29 [1] RSPM (R 4.3.0)
#>  evaluate      0.23    2023-11-01 [1] RSPM (R 4.3.0)
#>  fansi         1.0.6   2023-12-08 [1] RSPM (R 4.3.0)
#>  fastmap       1.1.1   2023-02-24 [1] RSPM (R 4.3.0)
#>  fs            1.6.3   2023-07-20 [1] RSPM (R 4.3.0)
#>  glue          1.7.0   2024-01-09 [1] RSPM (R 4.3.0)
#>  htmltools     0.5.7   2023-11-03 [1] RSPM (R 4.3.0)
#>  httr          1.4.7   2023-08-15 [1] RSPM (R 4.3.0)
#>  jquerylib     0.1.4   2021-04-26 [1] RSPM (R 4.3.0)
#>  jsonlite      1.8.8   2023-12-04 [1] RSPM (R 4.3.0)
#>  knitr         1.45    2023-10-30 [1] RSPM (R 4.3.0)
#>  lifecycle     1.0.4   2023-11-07 [1] RSPM (R 4.3.0)
#>  magrittr      2.0.3   2022-03-30 [1] RSPM (R 4.3.0)
#>  memoise       2.0.1   2021-11-26 [1] RSPM (R 4.3.0)
#>  pillar        1.9.0   2023-03-22 [1] RSPM (R 4.3.0)
#>  pkgconfig     2.0.3   2019-09-22 [1] RSPM (R 4.3.0)
#>  pkgdown       2.0.7   2022-12-14 [1] RSPM (R 4.3.0)
#>  purrr         1.0.2   2023-08-10 [1] RSPM (R 4.3.0)
#>  R6            2.5.1   2021-08-19 [1] RSPM (R 4.3.0)
#>  ragg          1.2.7   2023-12-11 [1] RSPM (R 4.3.0)
#>  rappdirs      0.3.3   2021-01-31 [1] RSPM (R 4.3.0)
#>  remotes       2.4.2.1 2023-07-18 [1] RSPM (R 4.3.0)
#>  rJava         1.0-11  2024-01-26 [1] RSPM (R 4.3.0)
#>  rlang         1.1.3   2024-01-10 [1] RSPM (R 4.3.0)
#>  rmarkdown     2.26    2024-03-05 [1] CRAN (R 4.3.2)
#>  rotemplate    2.0.0   2024-02-14 [1] Github (ropensci/rotemplate@864ebd0)
#>  rstudioapi    0.15.0  2023-07-07 [1] RSPM (R 4.3.0)
#>  sass          0.4.8   2023-12-06 [1] RSPM (R 4.3.0)
#>  sessioninfo   1.2.2   2021-12-06 [1] RSPM (R 4.3.0)
#>  stringi       1.8.3   2023-12-11 [1] RSPM (R 4.3.0)
#>  stringr       1.5.1   2023-11-14 [1] RSPM (R 4.3.0)
#>  systemfonts   1.0.5   2023-10-09 [1] RSPM (R 4.3.0)
#>  textshaping   0.3.7   2023-10-09 [1] RSPM (R 4.3.0)
#>  tibble        3.2.1   2023-03-20 [1] RSPM (R 4.3.0)
#>  utf8          1.2.4   2023-10-22 [1] RSPM (R 4.3.0)
#>  vctrs         0.6.5   2023-12-01 [1] RSPM (R 4.3.0)
#>  whisker       0.4.1   2022-12-05 [1] RSPM (R 4.3.0)
#>  withr         3.0.0   2024-01-16 [1] RSPM (R 4.3.0)
#>  xfun          0.42    2024-02-08 [1] RSPM (R 4.3.0)
#>  xml2          1.3.6   2023-12-04 [1] RSPM (R 4.3.0)
#>  yaml          2.3.8   2023-12-11 [1] RSPM (R 4.3.0)
#> 
#>  [1] /usr/local/lib/R/site-library
#>  [2] /usr/lib/R/site-library
#>  [3] /usr/lib/R/library
#> 
#> ──────────────────────────────────────────────────────────────────────────────

check_empty_beaustier_folders()